feat: Standalone UI overhaul - transport bar layout, crossfade slider, session registry, and visual polish
- Fix transport bar layout: centered fixed-width (365px), compact time label (85px), 140px progress bar
- Add crossfade slider with custom LookAndFeel for top-right corner
- Add BlindCardSessionRegistry for multi-instance session management
- Improve chip button, poker card, poker table, and star rating visuals
- Update question banner, quick start guide, and mode selector styling
- Fix CMake: disable LTO on Apple, force architecture and deployment target settings
- Add CLAUDE.md project instructions

+++ Source/UI/Components/PokerTable.cpp
@@ -83,7 +83,7 @@ PokerCard* PokerTable::getCard(int index)
 {
     if (index >= 0 && index < static_cast<int>(cards.size()))
     {
-        return cards[index].get();
+        return cards[static_cast<size_t>(index)].get();
     }
     return nullptr;
 }
@@ -531,7 +531,7 @@ void PokerTable::updateCardLayout()
 
         for (int i = 0; i < cardCount; ++i)
         {
-            cards[i]->setBounds(
+            cards[static_cast<size_t>(i)]->setBounds(
                 startX + i * (cardWidth + horizontalGap),
                 startY,
                 cardWidth,
@@ -557,7 +557,7 @@ void PokerTable::updateCardLayout()
 
         for (int i = 0; i < topRowCount; ++i)
         {
-            cards[i]->setBounds(
+            cards[static_cast<size_t>(i)]->setBounds(
                 topStartX + i * (cardWidth + horizontalGap),
                 topY,
                 cardWidth,
@@ -571,7 +571,7 @@ void PokerTable::updateCardLayout()
 
         for (int i = 0; i < bottomRowCount; ++i)
         {
-            cards[topRowCount + i]->setBounds(
+            cards[static_cast<size_t>(topRowCount + i)]->setBounds(
                 bottomStartX + i * (cardWidth + horizontalGap),
                 bottomY,
                 cardWidth,
